+++ apps/web/core/io/governance-space-plugin/abis/main-voting-plugin-abi.ts
