+++ apps/web/core/io/governance-space-plugin/abis/personal-space-admin-plugin-setup-abi.ts
