+++ apps/web/core/io/governance-space-plugin/abis/space-plugin-abi.ts
