+++ apps/web/core/io/governance-space-plugin/abis/space-plugin-setup-abi.ts
