+++ apps/web/core/io/governance-space-plugin/internal/interfaces.ts
